+++ arena.py
@@ -1,11 +1,12 @@
 import random
+import effects
+import powerups
 
 import numpy as np
-import effects
 
 from globals import GameInfo
 from camera import CameraState
-from util import Vector, get_main_path, painter_transform_with_rot, draw_img_with_rot
+from util import Vector, get_main_path, painter_transform_with_rot
 from constants import MAX_POWER_UP_ITER, TILES_PER_POWER_UP
 
 if not GameInfo.is_headless:
@@ -14,7 +15,6 @@ if not GameInfo.is_headless:
 
 tile_texture_path = get_main_path() + "/textures/static_tiles/"
 animated_tiles_texture_path = get_main_path() + "/textures/animated_tiles/"
-power_up_texture_path = get_main_path() + "/textures/power_ups/"
 
 
 class TileType:
@@ -64,77 +64,6 @@ tile_type_dict = {
 }
 
 
-class PowerUp:
-    def __init__(self, arena, effect, index, position):
-        self.power_up_type = type(self)
-        if self.power_up_type is PowerUp:
-            print("ERROR: PowerUp base class should not be instantiated!")
-
-        self.arena = arena
-        self.effect = effect
-        self.index = index
-        self.position = position
-        self.arena.power_up_list.append(self)
-        self._texture = None
-        self._texture_size = None
-
-    @property
-    def texture(self):
-        if self._texture is None:
-            self.load_image()
-        return self._texture
-
-    @property
-    def texture_size(self):
-        if self._texture_size is None:
-            self.load_image()
-        return self._texture_size
-
-    def load_image(self):
-        filename = power_up_texture_path + self.power_up_type.name + ".png"
-
-        self._texture = QPixmap(filename)
-        self._texture_size = Vector(self._texture.width(), self._texture.height())
-        if self._texture_size.x == 0 or self._texture_size.y == 0:
-            print("ERROR: texture for " + self.power_up_type.name
-                  + " power up has 0 size or is missing at " + filename + "!")
-
-    def apply(self, robot):
-        robot.effects.append(self.effect)
-        self.arena.power_ups[self.index.y][self.index.x] = None
-        self.arena.power_up_list.remove(self)
-
-    def draw(self, qp):
-        draw_img_with_rot(qp, self.texture, self.texture_size.x, self.texture_size.y, self.position, 0)
-
-
-class HealthPowerUp(PowerUp):
-    name = "health"
-    health_gain = 250
-
-    def __init__(self, arena, index, position):
-        effect = effects.HealthEffect(instant_change=HealthPowerUp.health_gain)
-        super().__init__(arena, effect, index, position)
-
-
-class SpeedPowerUp(PowerUp):
-    name = "speed"
-
-    def __init__(self, arena, index, position):
-        effect = effects.PowerUpSpeedEffect()
-        super().__init__(arena, effect, index, position)
-
-
-class DamagePowerUp(PowerUp):
-    name = "damage"
-    duration = 5
-    damage_factor = 2
-
-    def __init__(self, arena, index, position):
-        effect = effects.PowerUpDamageEffect(duration=DamagePowerUp.duration, damage_factor=DamagePowerUp.damage_factor)
-        super().__init__(arena, effect, index, position)
-
-
 class Arena:
     def __init__(self, tile_count):
         self.tile_count = tile_count.copy()
@@ -167,7 +96,7 @@ class Arena:
 
     def get_empty_power_ups(self):
         # get array of power_ups with correct dimensions
-        return np.empty((self.tile_count.y, self.tile_count.x), dtype=PowerUp)
+        return np.empty((self.tile_count.y, self.tile_count.x), dtype=powerups.PowerUp)
 
     # get different portal tiles for portal tile effects
     def get_portal_tiles(self):
@@ -182,7 +111,7 @@ class Arena:
         return self._portal_tiles
 
     # place power ups in the arena
-    def place_power_ups(self, delta_time):
+    def place_power_up(self, delta_time):
         if len(self.power_up_list) * TILES_PER_POWER_UP >= self.tile_count.x * self.tile_count.y:
             return
 
@@ -196,9 +125,9 @@ class Arena:
             if tile_type.name == "ground" and power_up is None:
                 position = Vector(x, y)
                 position.mult(GameInfo.arena_tile_size)
-                position.add(Vector(GameInfo.arena_tile_size / 2, GameInfo.arena_tile_size / 2))
+                position.add_scalar(GameInfo.arena_tile_size / 2)
                 # decide which type of power up to place
-                power_up_types = [HealthPowerUp, SpeedPowerUp, DamagePowerUp]
+                power_up_types = [powerups.HealthPowerUp, powerups.SpeedPowerUp, powerups.DamagePowerUp]
                 power_up_type = random.choice(power_up_types)
                 self.power_ups[y][x] = power_up_type(self, Vector(x, y), position)
                 break
