+++ constants.py
@@ -1,17 +1,18 @@
 from util import s_to_ns, Vector
 
 
-# Arena
-MAP_FORMAT_VERSION = 1
-MAX_POWER_UP_ITER = 10
-TILES_PER_POWER_UP = 100
-
 # Time
 FIXED_FPS = 60
 FIXED_DELTA_TIME = 1 / FIXED_FPS
 FIXED_DELTA_TIME_NS = s_to_ns(FIXED_DELTA_TIME)
 MAX_FIXED_TIMESTEPS = 3
 MAX_EXTRAPOLATION_STEPS = round((1 / 2) * FIXED_FPS)
+
+# Arena
+MAP_FORMAT_VERSION = 1
+MAX_POWER_UP_ITER = 10
+TILES_PER_POWER_UP = 100
+POWER_UPS_PER_S = round(5 / FIXED_DELTA_TIME)
 
 # Robot
 MAX_ROBOT_HEALTH = 1000
